Moved non-puzzle related code to own View files and refactored main puzzle generation page for shorter code
